@@ -8,17 +8,18 @@
   } from '../../utils/keyboard';
   import { createArray } from '../../utils/array';
   import { calculateScoreDelta } from '../../lib/wizard/game';
-  import type { Score } from '../../lib/wizard/types';
+  import type { Player, Score } from '../../lib/wizard/stores';
 
-  export let name: 'bids' | 'tricks';
-  export let round: number;
-  export let player: {
-    id: string;
-    name: string;
-    score: Score;
-  };
+  interface Props {
+    name: 'bids' | 'tricks';
+    round?: number;
+    player: Player;
+    score: Pick<Score, 'bid' | 'tricks'>;
+  }
 
-  $: delta = calculateScoreDelta(player.score);
+  const { name, round = 0, player, score = $bindable() }: Props = $props();
+
+  const delta = $derived(calculateScoreDelta(score));
 
   let timeoutId: ReturnType<typeof setTimeout>;
   let previousKeys = '';
@@ -32,22 +33,18 @@
   function handleKeyDown(event: KeyboardEvent) {
     if (isArrowUp(event) || isArrowRight(event)) {
       event.preventDefault();
-      player.score[key] = isNumber(player.score[key])
-        ? Math.min(max, player.score[key] + 1)
-        : 0;
+      score[key] = isNumber(score[key]) ? Math.min(max, score[key] + 1) : 0;
     }
     if (isArrowDown(event) || isArrowLeft(event)) {
       event.preventDefault();
-      player.score[key] = isNumber(player.score[key])
-        ? Math.max(min, player.score[key] - 1)
-        : 0;
+      score[key] = isNumber(score[key]) ? Math.max(min, score[key] - 1) : 0;
     }
     // Allow typing of the value, including numbers with multiple digits
     if (isNumber(Number.parseInt(event.key))) {
       clearTimeout(timeoutId);
       const number = Number.parseInt(`${previousKeys}${event.key}`);
       if (number >= min && number <= max) {
-        player.score[key] = number;
+        score[key] = number;
         previousKeys += event.key;
       }
       timeoutId = setTimeout(() => {
@@ -59,7 +56,7 @@
   }
 
   function handleSelect(event: MouseEvent) {
-    player.score[key] = Number.parseInt(
+    score[key] = Number.parseInt(
       (event.currentTarget as HTMLDivElement).dataset.value as string,
     );
   }
@@ -87,7 +84,7 @@
   class="input"
   type="number"
   {name}
-  bind:value={player.score[key]}
+  bind:value={score[key]}
   aria-labelledby={labelId}
   {min}
   {max}
@@ -100,18 +97,18 @@
   tabindex="0"
   aria-valuemin={min}
   aria-valuemax={max}
-  aria-valuenow={player.score[key]}
-  on:keydown={handleKeyDown}
+  aria-valuenow={score[key]}
+  onkeydown={handleKeyDown}
 >
   {#each createArray(round + 1) as _, option}
-    <!-- svelte-ignore a11y-click-events-have-key-events -->
-    <!-- svelte-ignore a11y-no-static-element-interactions -->
+    <!-- svelte-ignore a11y_click_events_have_key_events -->
+    <!-- svelte-ignore a11y_no_static_element_interactions -->
     <div
       class="option"
       data-value={option}
-      data-active={option === player.score[key]}
-      data-predicted={option === player.score.bid}
-      on:click={handleSelect}
+      data-active={option === score[key]}
+      data-predicted={option === score.bid}
+      onclick={handleSelect}
     >
       {option}
     </div>
@@ -123,10 +120,6 @@
     display: flex;
     gap: 1ch;
     align-items: baseline;
-  }
-
-  :global(.no-js) .delta {
-    display: none;
   }
 
   .positive {
@@ -155,17 +148,13 @@
     border: 1px solid var(--color-fg-default);
   }
 
-  :global(.no-js) .input {
-    display: block;
-  }
-
   .slider {
     display: inline-block;
-    border-radius: 1rem;
+    border-radius: 99999px;
   }
 
-  :global(.no-js) .slider {
-    display: none;
+  .slider:focus-visible {
+    outline: 2px solid var(--color-fg-subtle);
   }
 
   .option {
