@@ -1,29 +1,23 @@
-import { defineConfig } from 'astro/config';
-import db from '@astrojs/db';
+import { defineConfig, envField } from 'astro/config';
 import svelte from '@astrojs/svelte';
-import vercel from '@astrojs/vercel/serverless';
 import { browserslistToTargets } from 'lightningcss';
 import browserslist from 'browserslist';
 import AstroPWA from '@vite-pwa/astro';
 
 // https://astro.build/config
 export default defineConfig({
-  output: 'server',
-  adapter: vercel({
-    imageService: true,
-    functionPerRoute: false,
-  }),
-  security: {
-    checkOrigin: true,
-  },
   site: 'https://games.connorbaer.com',
-  prefetch: {
-    prefetchAll: true,
-    defaultStrategy: 'hover',
+  env: {
+    schema: {
+      LANGUAGE: envField.string({
+        context: 'client',
+        access: 'public',
+        default: 'en',
+      }),
+    },
   },
   integrations: [
     svelte(),
-    db(),
     AstroPWA({
       base: '/',
       scope: '/',
