@@ -1,9 +1,15 @@
+<script lang="ts">
+  import type { SVGAttributes } from 'svelte/elements';
+
+  const props: SVGAttributes<SVGSVGElement> = $props();
+</script>
+
 <svg
   xmlns="http://www.w3.org/2000/svg"
   width="24"
   height="24"
   viewBox="0 0 24 24"
-  {...$$props}
+  {...props}
 >
   <path
     fill="currentColor"
