@@ -1,10 +1,10 @@
-import { LOCALE } from '../lib/wizard/constants';
+import { LANGUAGE } from 'astro:env/client';
 
 export function capitalize(word: string): string {
   return word.charAt(0).toUpperCase() + word.slice(1);
 }
 
-const pluralRules = new Intl.PluralRules(LOCALE);
+const pluralRules = new Intl.PluralRules(LANGUAGE);
 
 export function pluralize(
   word: { singular: string; plural: string },
@@ -21,4 +21,13 @@ export function pluralize(
   }
 
   return word.plural;
+}
+
+const listFormat = new Intl.ListFormat(LANGUAGE, {
+  style: 'long',
+  type: 'conjunction',
+});
+
+export function join(list: string[]): string {
+  return listFormat.format(list);
 }
